pyproject.toml
[tool.poetry]
name = "kakeibo-exporter"
version = "0.1.0"
description = ""
authors = ["faruryo"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^1.8.0"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
sqlalchemy = "^1.3.16"
mysql-connector-python = "^8.0.19"
typer = "^0.2.1"
fastapi = "^0.54.1"
uvicorn = "^0.11.5"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
flake8 = "^3.7.9"
black = "^19.10b0"
isort = "^4.3.21"
mypy = "^0.770"
rope = "^0.16.0"
pytest-mock = "^3.1.0"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
kex = 'kakeibo_exporter.__main__:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

